updated presentation tree sample
